Moved tests that depends on enhetsregisteret to new integration test
