+++ applications/harvester-api/src/test/java/no/dcat/harvester/crawler/converters/PublisherOrganizationNumberAndNameTest.java
@@ -61,91 +61,5 @@ public class PublisherOrganizationNumberAndNameTest {
         Assert.assertThat(actual, Is.is(expected));
     }
 
-    @Test
-    public void organizationNumberFromGeonorgeForVegvesen() throws Throwable {
-        Model model = FileManager.get().loadModel("geonorge-data-2017-10-19.xml");
 
-        Resource publisherResource = model.getResource("https://register.geonorge.no/register/organisasjoner/kartverket/statens-vegvesen");
-
-        String previousName = publisherResource.getProperty(FOAF.name).getString();
-        logger.info("name before {}",previousName);
-
-        converter.collectFromModel(model);
-
-        String newName = publisherResource.getProperty(FOAF.name).getString();
-
-        OutputStream output2 = new ByteArrayOutputStream();
-        model.write(output2, "TURTLE");
-        logger.debug("[model_before_conversion] \n{}", output2.toString());
-
-        logger.info("name after {}",newName);
-
-        Assert.assertThat(newName, Is.is("STATENS VEGVESEN"));
-    }
-
-    @Test
-    public void organizationNumberInIdentifierUsesMasterRegistryNameOK() throws Throwable {
-        Model model = FileManager.get().loadModel("oljedir.xml");
-
-        Resource publisherResource = model.getResource("https://register.geonorge.no/register/organisasjoner/kartverket/oljedirektoratet");
-
-        String previousName = publisherResource.getProperty(FOAF.name).getString();
-        logger.info("name before {}",previousName);
-
-        converter.collectFromModel(model);
-
-        String newName = publisherResource.getProperty(FOAF.name).getString();
-
-        OutputStream output2 = new ByteArrayOutputStream();
-        model.write(output2, "TURTLE");
-        logger.debug("[model_before_conversion] \n{}", output2.toString());
-
-        logger.info("name after {}",newName);
-
-        Assert.assertThat(newName, Is.is("OLJEDIREKTORATET"));
-    }
-
-    @Test
-    public void organizationNameSubstitutionFromCanonicalTableOK() throws Throwable {
-        Model model = FileManager.get().loadModel("oljedir.xml");
-
-        Resource publisherResource = model.getResource("https://register.geonorge.no/register/organisasjoner/kartverket/difi");
-
-        String previousName = publisherResource.getProperty(FOAF.name).getString();
-        logger.info("name before {}",previousName);
-
-        converter.collectFromModel(model);
-
-        String newName = publisherResource.getProperty(FOAF.name).getString();
-
-        OutputStream output2 = new ByteArrayOutputStream();
-        model.write(output2, "TURTLE");
-        logger.debug("[model_before_conversion] \n{}", output2.toString());
-
-        logger.info("name after {}",newName);
-
-        Assert.assertThat(newName, Is.is("DIFI"));
-    }
-
-    @Test
-    public void organizationNameSubstitutionFromCanonicalTableWhenOfficalURIIsUsedOK() throws Throwable {
-        Model model = FileManager.get().loadModel("oljedir.xml");
-
-        Resource publisherResource = model.getResource("http://data.brreg.no/enhetsregisteret/enhet/971526920");
-
-        String previousName = publisherResource.getProperty(FOAF.name).getString();
-        logger.info("name before {}",previousName);
-
-        converter.collectFromModel(model);
-
-        String newName = publisherResource.getProperty(FOAF.name).getString();
-
-        OutputStream output2 = new ByteArrayOutputStream();
-        model.write(output2, "TURTLE");
-        logger.debug("[model_before_conversion] \n{}", output2.toString());
-
-        logger.info("name after {}",newName);
-
-        Assert.assertThat(newName, Is.is("SSB"));
-    }
 }
